% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{COmet}
\alias{COmet}
\title{Estimate the COMET using adaptive thresholding}
\usage{
COmet(dat, lambda = NULL, N = 100, mul = 2, tol = 1e-06)
}
\arguments{
\item{dat}{n by p matrix of multivariate data}

\item{lambda}{sequence of threshold values for universal thresholding}

\item{N}{number of intervals for adaptive thresholding}

\item{mul}{maximum level of threshold (times of 1 standard error of each sample covariance) for adaptive thresholding}

\item{tol}{tolerance level of errors}
}
\value{
cov_list: list of covariance matrix estimators for all thresholds

aic: sequence of the values of AIC for all thresholds

bic: sequence of the values of BIC for all thresholds
}
\description{
Estimate the COMET using adaptive thresholding
}
\examples{
library(mvtnorm)
p <-10; n<- 100
Sigma = diag(p); diag(Sigma[-p,-1])=0.5; diag(Sigma[-1,-p])=0.5
dat <- rmvnorm(n,mean=rep(0,ncol(Sigma)),sigma=Sigma)
res_uni = COmet(dat, lambda = seq(0,1,0.01))
res_uni$cov_list[[which.min(res_uni$aic)]]; res_uni$cov_list[[which.min(res_uni$bic)]]
res_ada = COmet(dat, mul = 3)
res_ada$cov_list[[which.min(res_ada$aic)]]; res_ada$cov_list[[which.min(res_ada$bic)]]
}
