% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cometAda}
\alias{cometAda}
\title{Estimate the COMET using adaptive thresholding}
\usage{
cometAda(dat, N = 100, mul = 2, tol = 1e-06)
}
\arguments{
\item{dat}{n by p matrix of multivariate data}

\item{N}{number of intervals}

\item{mul}{maximum level of threshold (times of 1 standard error of each sample covariance)}

\item{tol}{tolerance level of errors}
}
\value{
cov_list1: list of covariance matrix estimators for all thresholds

aic1: sequence of the values of AIC for all thresholds

bic1: sequence of the values of BIC for all thresholds
}
\description{
Estimate the COMET using adaptive thresholding
}
\examples{
x = matrix(rnorm(100*10, 0, 1),100,10)
}
