% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{covchaud}
\alias{covchaud}
\title{Perform iterative conditional fitting (ICF) starting from the sample covariance matrix}
\usage{
covchaud(amat, dat, cc = 0, tol = 1e-06)
}
\arguments{
\item{amat}{p by p adjacency matrix for indicating the location of non-zero covariances}

\item{dat}{n by p matrix of multivariate data}

\item{cc}{add diagonal if the sample covariance matrix is not positive-definite}

\item{tol}{tolerance level of errors}
}
\value{
mat: ICF estimator
}
\description{
Perform iterative conditional fitting (ICF) starting from the sample covariance matrix
}
\examples{
library(mvtnorm)
p <-10; n<- 100
Sigma = diag(p); diag(Sigma[-p,-1])=0.5; diag(Sigma[-1,-p])=0.5
dat <- rmvnorm(n,mean=rep(0,ncol(Sigma)),sigma=Sigma)
amat = (Sigma!=0)*1
covchaud(amat, dat)
}
